Add BaseOtlpExporterOptions
