+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/.publicApi/netstandard2.1/PublicAPI.Unshipped.txt
@@ -1,0 +1,12 @@
+OpenTelemetry.Exporter.BaseOtlpExporterOptions
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.BaseOtlpExporterOptions(Microsoft.Extensions.Configuration.IConfiguration configuration) -> void
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Endpoint.get -> System.Uri
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Endpoint.set -> void
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Headers.get -> string
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Headers.set -> void
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.HttpClientFactory.get -> System.Func<System.Net.Http.HttpClient>
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.HttpClientFactory.set -> void
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Protocol.get -> OpenTelemetry.Exporter.OtlpExportProtocol
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.Protocol.set -> void
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.TimeoutMilliseconds.get -> int
+OpenTelemetry.Exporter.BaseOtlpExporterOptions.TimeoutMilliseconds.set -> void
